первый вариант реализации
